Move peamu in a new allauth folder
